GroupSort is now TPU compatible
Dynamic reshape cannot be performed on TPU, the intermediate shape
and output shape are stored during layer build.

Note that output shape is equal to input shape, except that the None
batch size must be replaced with -1, to be able to reshape it at the end
of the call. For example:
	input_shape = (None, H, W, C)
	output_shape = (-1, H, W, C) (required for tf.reshape)