chore: VERSION file to have it at a single place
The aim of this commit is to have the package version at a single place, and
not across multiple files.

The VERSION file contains the version number. It is now the only place where
it is set. The setup.py and the Sphinx conf.py read this file.
It is now possible to get the version number directly from the package:
  import deel.lip
  print(deel.lip.__version__)

Note that to add non-Python files to the sdist package, it is required to add
`include_package_data=True` in setup.py and a MANIFEST.in file containing the
non-Python files to add in package.