feat: centered hinge losses have now margin/2 on both sides
The idea is to enforce a margin m between the true logit and other
logits. Since the hinge losses enforce to be centered around zero, we want to
have the true logit above m/2 and other logits below -m/2, then a margin m
between the true logit and others.

Unit tests have been modified in order to reflect this change: min_margin value
must be doubled.