feat: if condition in HKR losses moved to __init__
For now, an if condition depending on alpha was implemented in the `call()`
function of the HKR and MulticlassHKR losses. Moreover, the hyper-parameter
alpha is now a tf.Variable which can be modified during training (and then
the if condition can be misinterpreted).

To avoid hidden problems and to handle TPU training, the if condition is moved
to the `__init__()`.
Note a consequence of this commit is that the `self.fct` is defined at
initialization of the loss and we cannot pass during training from an alpha
np.inf to a float alpha, and vice versa.