fix(test): random test was failing. Fix better seed.
The power iteration test for Dense was sometimes failing due to a random
choice of kernel and initialization vector u. Even if np.random.seed(42)
was set at the beginning of the file.

To ensure a fully deterministic test (deterministic kernel and
initialization u), a random generator np.random.default_rng(42) was created
and used to create both kernel and u.
The results for Dense test are then deterministic, either for a single test
file or for the whole unittest. Note that this is the only test modified;
other tests of the file pass, they have not been changed with deterministic
initializations.