fix(compute_layer_sv): remove bias and activation in convolution SVD
When building the equivalent matrix of a convolution layer, the bias and
activation must be removed to only keep the convolutional operation of the
layer.
By cloning the layer and setting use_bias and activation to False and None
respectively, only the convolution is performed.
Note that cloning the model already built the new model with the correct
input shape.