feat(normalizers): upgrade to Keras 3
In Keras 3, `swap_memory` and `parallel_iterations` are not arguments
in while_loop().

In Keras 3, `l2_normalize()` does not exist. We recreate one in utils.py
based on TensorFlow implementation.

In Keras 3, `keras.ops.norm()` can only be applied on 1-D or 2-D tensors.
In the power iteration method, we flatten `u` to compute the stopping
condition.