feat(layers): save/load own variables in dense and conv
In Keras 3, when model.save() and load_model(), custom variables are
not saved by default. It is necessary to override save_own_variables() and
load_own_variables() to manually store custom variables.

Note that Conv2DTranspose has not the same behaviour: all trainable and
non-trainable variables are saved/loaded. It is thus not required to override
these two methods.