feat: change Lipschitz cst estimation to Jacobian method
To estimate the Lipschitz constant of a model, we now use the Jacobian
method that is more robust than random perturbations (subject to numerical
errors).

The Jacobian method computes Jacobians for a batch of inputs and gets
the max singular values of each Jacobian matrix, corresponding to the local
Lipschitz constant.
The function then returns the greatest Lipschitz constant estimation in the
batch.